services/product-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-decouple
psycopg2-binary
drf-spectacular
django-cors-headers
